Cleaned up DB code by moving multiply code into a private class (#19)
Cleaned up DB code by moving the multiply into a private class. This allows it to track the database iterator internally, and move a lot of the recursive function parameters to class members. That makes it a lot easier to understand and work with. It's now more obvious which parameters are modified during recursion and which are simply context.